requirements.txt
asgiref==3.3.4
autopep8==1.5.6
Django==3.2
django-js-asset==1.2.2
django-mptt==0.12.0
django-simple-bulma==2.2.0
djangorestframework==3.12.4
libsass==0.20.1
pycodestyle==2.7.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
